Fix: filtering products by category
